Gerhard brings MSVC up to speed. Resync GPSBabel.pro Kill mkwintesto Various MSVC weirdness. A change in main to set character encoding that really doesn't make sense to me since he changed only the <Qt5 case and he's on Qt5.

+++ gpsbabel/waypt.cc
@@ -595,7 +595,7 @@ waypt_speed_ex(const waypoint* A, const waypoint* B)
     return 0;
   }
 
-  time = fabs(A->creation_time.msecsTo(B->creation_time)) / 1000.0;
+  time = fabs((double)A->creation_time.msecsTo(B->creation_time)) / 1000.0;
   if (time > 0) {
     return (dist / time);
   } else {
@@ -618,7 +618,7 @@ waypt_speed(const waypoint* A, const waypoint* B)
     return 0;
   }
 
-  time = fabs(A->creation_time.msecsTo(B->creation_time)) / 1000.0;
+  time = fabs((double)A->creation_time.msecsTo(B->creation_time)) / 1000.0;
   if (time > 0) {
     return (dist / time);
   } else {
